+++ axelor-account/src/main/java/com/axelor/apps/account/db/repo/PaymentSessionAccountRepository.java
@@ -19,7 +19,7 @@ package com.axelor.apps.account.db.repo;
 
 import com.axelor.apps.account.db.PaymentSession;
 import com.axelor.apps.account.exception.AccountExceptionMessage;
-import com.axelor.apps.account.service.PaymentSessionService;
+import com.axelor.apps.account.service.payment.paymentsession.PaymentSessionService;
 import com.axelor.apps.base.db.Company;
 import com.axelor.apps.base.db.repo.SequenceRepository;
 import com.axelor.apps.base.service.administration.SequenceService;
