+++ axelor-account/src/main/java/com/axelor/apps/account/service/payment/paymentsession/PaymentSessionService.java
@@ -15,7 +15,7 @@
  * You should have received a copy of the GNU Affero General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-package com.axelor.apps.account.service;
+package com.axelor.apps.account.service.payment.paymentsession;
 
 import com.axelor.apps.account.db.Journal;
 import com.axelor.apps.account.db.PaymentSession;
